Moved bowtie generation to within report fcn
Added parameter to Write-PISysAuditReport to accept ComputerParamsTable
so this may be passed to Write-PISysAudit_BowTieDiagrams.

Changed return value of Write-PISysAudit_BowTieDiagrams to be a
hashtable keyed on the machine names. Each machine key has a value with
an array of report filenames. This will be useful for inserting bowtie
diagrams into the HTML report since it allows easier sorting.