Moved bowtie template generation to main module
Added New-PISysAudit_BowTieTemplate in core library. This function will
read the JSON definition files and build the SVG template for the bow
ties each time the module is run. This means that changes to the bowtie
need only be made in the definition file and they will be reflected in
the audit output.

Also added Audit Item ID to bowtie, updated comments, and other
miscellaneous fixes.